requirements.txt
python-telegram-bot==12.7
PyPDF2==1.26.0